Moved add logic to within component
