+++ client/src/components/dashboard/visas/addVisa.js
@@ -1,3 +1,5 @@
+import React, { useState } from "react";
+import axios from "axios";
 import Button from "@mui/material/Button";
 import TextField from "@mui/material/TextField";
 import Dialog from "@mui/material/Dialog";
@@ -8,14 +10,62 @@ import DialogTitle from "@mui/material/DialogTitle";
 import Fab from "@mui/material/Fab";
 import AddIcon from '@mui/icons-material/Add';
 
-export default function AddVisa({
-  open,
-  handleOpen,
-  handleClose,
-  handleChange,
-  visa,
-  handleSubmit,
-}) {
+const AddVisa = (props) => {
+  const userId = props.userId;
+  const tripId = props.tripId;
+  const visaId = props.visaId;
+  const open = props.open;
+  const handleOpen = props.handleOpen;
+  const handleClose = props.handleClose;
+  const visaData = props.visaData;
+  const [visa, setVisa] = useState({
+    visaNumber: visaData.visaNumber,
+    startDate: visaData.startDate,
+    endDate: visaData.endDate,
+    issuingCountry: visaData.issuingCountry,
+    user: userId,
+    trip: tripId,
+  });
+
+  const handleChange = (e) => {
+    const value = e.target.value;
+    setVisa({
+      ...visa,
+      [e.target.name]: value,
+    });
+  };
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+
+    const { visaNumber, startDate, endDate, issuingCountry, user, trip } =
+      visa;
+
+    const newVisa = {
+      visaNumber,
+      startDate,
+      endDate,
+      issuingCountry,
+      user: userId,
+      trip: tripId,
+    };
+
+    await axios
+      .post("http://localhost:8000/dashboard/visas/", newVisa)
+      .catch((err) => console.log(err.message))
+      .then(() => {
+        setVisa({
+          visaNumber: "",
+          startDate: "",
+          endDate: "",
+          issuingCountry: "",
+          user: userId,
+          trip: tripId,
+        });
+        handleClose();
+      });
+  };
+
   return (
     <div>
       <Fab size="large" color="secondary" aria-label="add" onClick={handleOpen}>
@@ -90,3 +140,5 @@ export default function AddVisa({
     </div>
   );
 }
+
+export default AddVisa;
